(refactor) updated API reference
