+++ src/applications/check-in/day-of/pages/Landing.jsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 
 import recordEvent from 'platform/monitoring/record-event';
 
 import { getTokenFromLocation, URLS, goToNextPage } from '../utils/navigation';
-import { api } from '../api';
+import { api } from '../../api';
 import { tokenWasValidated, triggerRefresh } from '../actions';
 import { setCurrentToken, clearCurrentSession } from '../../utils/session';
 import { createAnalyticsSlug } from '../../utils/analytics';
@@ -76,7 +75,7 @@ const Landing = props => {
   );
   return (
     <>
-      <LoadingIndicator message={loadMessage} />
+      <va-loading-indicator message={loadMessage} />
     </>
   );
 };
