+++ src/applications/check-in/pre-check-in/pages/EmergencyContact/index.jsx
@@ -7,7 +7,7 @@ import recordEvent from 'platform/monitoring/record-event';
 
 import { recordAnswer } from '../../actions';
 
-import { api } from '../../api/';
+import { api } from '../../../api';
 
 import BackButton from '../../components/BackButton';
 import BackToHome from '../../components/BackToHome';
