+++ src/applications/check-in/pre-check-in/pages/Landing/index.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import recordEvent from 'platform/monitoring/record-event';
 
-import { api } from '../../api';
+import { api } from '../../../api';
 
 import { createInitFormAction, createSetSession } from '../../actions';
